requirements.txt
Django>=1.4
pyquery
html2text
django-taggit
sorl-thumbnail
bambu-markup>=2.0.0
bambu-oembed>=2.0.0
bambu-attachments>=2.0.0
bambu-xmlrpc>=2.0